# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Notes Service
  version: "1.0"

servers:
  - url: http://127.0.0.1:8000

components:
  schemas:
    error:
      type: object
      properties:
        error:
          type: string
    note:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        title:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time
    notes:
      type: object
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/note"

  responses:
    internalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
    badRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
    notFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"

paths:
  /note:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
        required: true
      responses:
        "201":
          description: Заявка создана
        "400":
          $ref: "#/components/responses/badRequest"
        "500":
          $ref: "#/components/responses/internalServerError"
  /note/{id}:
    get:
      parameters:
        - name: id
          in: path
          schema:
            type: integer
            format: int64
            minimum: 1
          required: true
      responses:
        "200":
          description: Пользователь получает заявку
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/note"
        "404":
          $ref: "#/components/responses/notFound"
        "500":
          $ref: "#/components/responses/internalServerError"
  /notes:
    get:
      responses:
        "200":
          description: Пользователь получает список заявок
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notes"
        "500":
          $ref: "#/components/responses/internalServerError"